Relocated data_grid and search_bar folders
