+++ plugins/main/public/components/common/data_grid/use_data_grid.ts
@@ -3,7 +3,7 @@ import React, { useEffect, useMemo, useState, Fragment } from "react";
 import { SearchResponse } from "@opensearch-project/opensearch/api/types";
 // ToDo: check how create this methods
 import { parseData, getFieldFormatted, parseColumns } from './data_grid_service';
-import { IndexPattern } from '../../../../../../../src/plugins/data/common';
+import { IndexPattern } from '../../../../../../src/plugins/data/common';
 
 const MAX_ENTRIES_PER_QUERY = 10000;
