+++ plugins/main/public/components/common/search_bar/index.ts
